Big change for double time format

+++ src/main/java/com/chess/dto/rest/request/CreateGameRequestDTO.java
@@ -20,7 +20,7 @@ public class CreateGameRequestDTO {
     private String gameType = "standard";
     
     // Optional field for time control
-    private Integer timeControlMinutes;
+    private Double timeControlMinutes;
     
     // Optional field for game settings
     private Boolean isRated = false;
